Reorganized files to match master's, added checks for ratings, implemented requested changes and revisions
