+++ app/api/contributors/index.js
@@ -46,10 +46,10 @@ router.get('/', restricted, async (req, res) => {
   `, { requestedSkills: skills });
 
   const users = (await query).rows;
-  
-  var i;
-  for(i = 0; i < users.length; i++){
-    users[i].average_rating = users[i].total_stars/users[i].total_reviews;
+
+  let i;
+  for (i = 0; i < users.length; i += 1) {
+    users[i].average_rating = users[i].total_stars / users[i].total_reviews;
     delete users[i].total_reviews;
     delete users[i].total_stars;
   }
