+++ app/api/users/index.js
@@ -31,7 +31,7 @@ router.get('/', restricted, async (req, res) => {
   }
 });
 
-router.get('/:id'/*, restricted*/, async (req, res) => {
+router.get('/:id'/* , restricted */, async (req, res) => {
   const { id } = req.params;
 
   try {
@@ -58,7 +58,7 @@ router.get('/:id'/*, restricted*/, async (req, res) => {
           .json({ message: 'This account has been deactivated' });
       }
     }
-    user["average_rating"] = user.total_stars/user.total_reviews;
+    user.average_rating = user.total_stars / user.total_reviews;
     return res.status(200).json({ user, message: 'The user was found' });
   } catch (err) {
     log.error(err);
@@ -85,8 +85,8 @@ router.get('/sub/:sub', restricted, async (req, res) => {
           .status(401)
           .json({ message: 'This account has been deactivated' });
       }
-      
-      user.average_rating = user.total_stars/user.total_reviews;
+
+      user.average_rating = user.total_stars / user.total_reviews;
       delete user.total_stars;
       delete user.total_reviews;
       return res.status(200).json({ user, message: 'The user was found' });
