+++ knexfile.js
@@ -24,9 +24,9 @@ module.exports = {
     },
     pool: {
       min: 2,
-      max: 10
+      max: 10,
     },
-    useNullAsDefault: true
+    useNullAsDefault: true,
   },
 
   production: {
